Update Mutt config
Now with support for multiple accounts, invoke mutt by calling one of
the account-specific configurations, which source the master .muttrc
file.

This will allow for common usage across several accounts and prevent a
web broswer from being required to access infrequent messages.

This will be tied in with tmux, so that several windows will run with
different instances of mutt.