Add non-blocking option to pdflatex function
Prior to this change, any issue with syntax or format in the watched
pdflatex file would cause the build loop to await input (so that the
build could be cancelled).

This change will instead instruct the process to complete, passing over
the failure and doing 'best-effort' to generate the document.